editoast: move train schedule schedule item to schemas
