+++ editoast/editoast_schemas/src/train_schedule.rs
@@ -2,6 +2,9 @@ mod margins;
 pub use margins::MarginValue;
 pub use margins::Margins;
 
+mod schedule_item;
+pub use schedule_item::ScheduleItem;
+
 mod allowance;
 pub use allowance::Allowance;
 pub use allowance::AllowanceDistribution;
@@ -15,6 +18,7 @@ pub use rjs_power_restriction_range::RjsPowerRestrictionRange;
 
 editoast_common::schemas! {
     margins::schemas(),
+    schedule_item::schemas(),
     // TODO TrainSchedule V1 (it will be removed)
     allowance::schemas(),
     rjs_power_restriction_range::schemas(),
