+++ editoast/src/modelsv2/train_schedule.rs
@@ -2,12 +2,12 @@ use chrono::DateTime;
 use chrono::Utc;
 use editoast_derive::ModelV2;
 use editoast_schemas::train_schedule::Margins;
+use editoast_schemas::train_schedule::ScheduleItem;
 
 use crate::schema::v2::trainschedule::Comfort;
 use crate::schema::v2::trainschedule::Distribution;
 use crate::schema::v2::trainschedule::PathItem;
 use crate::schema::v2::trainschedule::PowerRestrictionItem;
-use crate::schema::v2::trainschedule::ScheduleItem;
 use crate::schema::v2::trainschedule::TrainScheduleOptions;
 
 #[derive(Debug, Default, Clone, ModelV2)]
